+++ apps/api/src/app/app.module.ts
@@ -1,11 +1,19 @@
 import { Module } from "@nestjs/common";
+import { LoggerModule } from "nestjs-pino";
+import { loggerConfig } from "@/config/logger.config";
 import { PrismaModule } from "@/infrastructure/prisma/prisma.module";
+import { HealthModule } from "@/modules/health/health.module";
 import { UsersModule } from "@/modules/users/users.module";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
 
 @Module({
-    imports: [PrismaModule, UsersModule],
+    imports: [
+        LoggerModule.forRoot(loggerConfig),
+        PrismaModule,
+        UsersModule,
+        HealthModule,
+    ],
     controllers: [AppController],
     providers: [AppService],
 })